Stop counting rewards as part of my stake summary
After pulling out pending rewards from the fully
protected and claimable amount in the table, in
order to make the sum of displayed positions match
the summary, it is also necessay to not count
pending rewards as part of the staked summary.

This also makes sense because pending rewards are
actually not staked, not accruing fees or extra
rewards, so not part of any staked position, so
imo good to not count them in a stake summary.